:bookmark: SCHISM v5.6.1 rev. 4997
SCHISM v5.6.1 is the version used for various papers by Khan et al.,
especially Khan et al. 2018. This version is often used as benchmark for
Khan et al., hence requires preserving in personal repository for future
use.

The original reposity is http://columbia.vims.edu/schism/tags/v5.6.1
which now seems to be down as SCHISM has moved fully to the github
repository. I did not realized that the history in the svn repo is lost
when moved to github repository, as there are tools to convert the old
histories, from svn to git.